test/requirements.txt
azure-storage-blob
beautifulsoup4
